requirements.txt
numpy
keyboard
Pillow
pypiwin32
scikit-learn
PyAutoGUI
opencv-python
ultralytics